Add Alek's version of libc and libm.
While I would like to use GNU's libc and libm, I feel like they would be too
"Linux-specific" - meaning the libc/libm would have some hard dependencies on
features provided by the Linux kernel which may or may not be provided by ALEK.

I took the high road on this one, and I am writing my own flavor of libc and
libm. This commit moved a lot of things around (e.g. splitting up the contents
of the top-level include/ directory between core/ and libc), as well as lots of
Makefile messing. I haven't tried to compile this for ARM version, so things
may be broken as we start to move things into the main branch.